Update and rename Maksim_Panibrat_HW7_Functions_Realizations.cpp to Maksim_Panibrat_HW7_Test_Functions_Realizations.cpp
